[ROCm] hermite_polynomial_h_forward workaround
Large n input caused a regression starting in ROCm 6.1. The for loop
will run for an excessive number of iterations. The root cause seems to
be how static_cast<int64_t> behaves for large float values such as 1e20
that certain unit tests will use. The workaround is to break out of the
loop once the returned value reaches nan.